원격(a34bdfb, hantawon123)의 identifier int 변환(api-server user_id/category_id/
tag_id/history_id/token_id/image_id + event-contract imageId/userId)을 채택하고,
로컬에서 진행한 analysis-worker job_id → integer 변환을 함께 유지한다.

충돌 해결:
- AnalysisResultRow.java: jobId Integer(로컬) + imageId Integer(양쪽 동일) 유지
- ImageAnalysisConsumer.java: event-contract가 이제 int를 직접 주므로 원격의
  `Integer imageId = payload.imageId()` 채택(로컬의 수동 파싱 제거)
- analysis-worker 001-init-schema.sql: job_id 전용 changeset으로 축소
  (image_id는 원격의 002-convert-image-id-to-integer.sql이 이미 처리하므로
  중복 제거, 팀 컨벤션대로 precondition 가드 추가)

+++ backend/api-server/src/main/java/com/ssafy/modera/api/domain/user/controller/AuthController.java
@@ -91,7 +91,7 @@ public class AuthController {
     })
     @PostMapping("/logout")
     public ResponseEntity<ApiResponse<LogoutResponse>> logout(
-            @AuthenticationPrincipal Long userId,
+            @AuthenticationPrincipal Integer userId,
             @RequestBody @Valid LogoutRequest request
     ) {
         return ResponseEntity.ok(ApiResponse.success(authService.logout(userId, request)));
