+++ backend/api-server/src/main/java/com/ssafy/modera/api/domain/user/repository/RefreshTokenRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Integer> {
 
-    Optional<RefreshToken> findByUserIdAndDeviceId(Long userId, String deviceId);
+    Optional<RefreshToken> findByUserIdAndDeviceId(Integer userId, String deviceId);
 
     Optional<RefreshToken> findByTokenHash(String tokenHash);
 }
